Refactor server structure using wrapper functions
* Update src/server.ts to contain only server definitions (createServer,
  wrapper functions, types)
* Create src/index.ts to be the entry point that creates server, define
  routes and make the server listen to requests
* Create src/controllers/ folder to store HTTP controllers
