+++ src/controllers/status.controller.ts
@@ -1,0 +1,5 @@
+import { defineController } from "../server.js";
+
+export const StatusController = defineController(http => {
+  http.get('/v1/status', () => ({ ok: true }));
+});
